Some services are implemented

+++ src/main/java/dev/serhats/hoodie/service/UserService.java
@@ -15,5 +15,5 @@ public interface UserService extends UserDetailsService {
 
     String getAccessToken(UserLogin userLogin);
 
-    UserView getUserById(int userId);
+    UserView getUserById(long userId);
 }
